+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 12.8
 - Display Enrollment Calendar to Student in Calendar.php
 - CSS add .calendar-block class in CalendarDay.inc.php & stylesheet.css
 - CSS themes transition slide menu & help in rtl.css, stylesheet.css & zresponsive.css
+- JS slide menu & help, use CSS class .slide in warewhouse.js
 
 Changes in 12.7.3
 -----------------